Moved Context and PaginationOutputModel to folder "types" and removed interface IRepo (since it was not being so useful)
